Fixed broken tests, tests on windows notifications
